sharding.cc: Correct integer conversion of YAML scalar
It turns out `.as<uint8_t>()` converts a scalar to the encoded value of the
first character of the string, rather than interpreting the string as an
integer.  For example, "0" is converted to 48 instead of 0.

+++ src/sharding.cc
@@ -37,7 +37,7 @@ ShardingInfo::ShardingInfo(const std::string& filename,
     auto failure_domain = config_backend["failure-domain"];
     if (failure_domain.IsDefined()) {
       KJ_REQUIRE(failure_domain.IsScalar());
-      backend.failure_domain = failure_domain.as<uint8_t>();
+      backend.failure_domain = failure_domain.as<int>();
     }
 
     KJ_CONTEXT(backend.addr, backend.failure_domain);
